pyproject.toml
[project]
name = "tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.1",
    "openai-agents>=0.2.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
